Define CUDA_GVDB_COPYDATA_PTX and CUDA_GVDB_MODULE_PTX macros for "cuda_gvdb_copydata.ptx" and "cuda_gvdb_module_ptx", based on notes from Issue #79 by icoderaven.
Here are some of the design considerations that went into this:

* I like this because defining it as a macro means that autocomplete works correctly,
and it's easier to change the names of the files in the future if ever needed.

* The source version (icoderaven@8e0cf00) defines the full path to the file. However,
we'd like it if it were possible to redistribute programs built using GVDB - e.g.
build on one computer and run on another, or move applications from one directory
to another - so we keep the relative paths. This does mean that if the executable
is moved, then the .ptx files have to be moved with it, but I think this makes sense.

* I moved the section where we add these compile definitions to near the bottom
of CMakeLists.txt, just before calling export(), near the rest of the calls
to target_compile_definitions.