Fix issue #83: Point filter mode, SetSteps and lights in world-space, larger epsilon.
Longer summary:

- Fixed a filter mode reversion that occurred as a result of switching from
texture and surface references to texture and surface objects in GVDB 1.1.1.
By default, texture and surface references used point filtering with a clamped
address mode, but I changed it to use linear filtering and wrapping in 1.1.1.
Unfortunately, this makes some kernels produce incorrect results if the filter
and addressing modes haven't been overwritten - in this case, kernelCopyTexZYX
sampled at voxel corners instead of the centers of voxels, and as a result
blended together voxels (including ones on opposite sides of the same brick!)
when transferring from an OpenVDB brick to a GVDB brick in LoadVDB. This was
responsible for some gridline artifacts in the resulting volume.

- It turns out that SetSteps uses application-space units (i.e. before mInvXform
transforms objects to index space) instead of index-space units when telling
GVDB how often to sample voxel volumes. However, a recent GVDB 1.1.1 commit made
it so that GVDB loaded OpenVDB transforms correctly (instead of using a constant
voxelsize of (1,1,1)). Since gImportVDB uses a model with a scale factor of
1/20, this meant that the model changed from being sampled about 4 times per
voxel to being sampled about once every 5 voxels (or less), resulting in visible
undersampling artifacts that looked like cross sections.

- However, it also turned out that lights in GVDB were provided in index space
instead of world-space. The API has been changed to take lights in world-space
instead to fix this inconsistency, and to transform light positions to
index-space when preparing to ray trace.

- Ccommit 07fa2811 from GVDB 1.1 changed the default epsilon to 0.001
from 0.01. However, this introduced artifacts in the gImportVDB sample, as the
volume was large enough for floating-point error to become a factor in the
higher-level HDDA steps. To fix this, the sample now changes its epsilon back to
0.01.

- Finally, this commit adds some fixes for OpenVDB 7.0 in the current version of
vcpkg. It looks like either vcpkg or OpenVDB's build system now gives the debug
library and DLL the names `openvdb_d.lib` and `openvdb_d.dll`, respectively, but
doesn't update the reference in `openvdb_d.lib` to the DLL. As a result,
openvdb_d.lib thinks that it needs `openvdb.dll` (not `openvdb_d.dll`), which
causes an error at runtime. As I expect this will be fixed in vcpkg or in
OpenVDB in the future, I've added a workaround step in the README file.