Revise VDB loading/saving logic, fixing issue #84. Removes mOVDB.
dvicini's note that the SaveVDB logic was producing volumes that looked corrupt
wound up turning up several bugs, which this commit fixes:

- kernelRetrieveTexXYZ rejected threads with indices larger than the atlas size,
instead of the brick size (the latter is correct, as the kernel only operates
on a single brick). Because the kernel launch was also running a larger grid than
necessary, this wound up copying adjacent bricks into the buffer's memory.

- LoadVDB supports both <3, 3, 3, 4> and <5, 4, 3> OpenVDB configurations, but
SaveVDB only supported brick sizes with a log dim of 3 (i.e. <.., .., 3>). This
fixes that by templating the core of SaveVDB, so that adding instantiations for
different OpenVDB grids is easy. Along the way, I also templated SaveVDB; this
made it so that I could remove the decision trees on grid types.

- SaveVDB would crash if mOVDB wasn't created. I was able to remove mOVDB entirely,
which allowed me to replace vdbCheck, vdbOrigin, and vdbNext.

This commit also changes AtlasRetrieveTexXYZ and AtlasCopyTexZYX to run fewer
blocks when the block size divides the brick resolution. It also adds
header documentation for ConvertToScalar, LoadVDB, and SaveVDB. Finally, it changes
the different type definitions in gvdb_volume_gvdb.h to use `using` instead of
`typedef`. LoadVDBInternal and SaveVDBInternal are listed in gvdb_volume_gvdb.h,
but are protected, as they are intended to be used inside Load/SaveVDB.

Inside gvdb_volume_gvdb, it also