Fixes a memory leak in gFluidSurface when pressing backtick and then 2, and other cleanup. Thanks to Patricio Astudillo for pointing this out.
The bug was that if you turned off the GUI, but enabled the option to draw a
point for each of the particles, the debug draw system would queue up more and
more lines to draw every frame until it ran out of memory. The quick fix in this
case was to make it so that these are only drawn when the GUI is visible.

- Removes mPackBuf and mPackGrid from fluid_system.cpp, as the sample used to
allocate these, not use them, and then leak the memory. Also removes the unused
functions FluidSystem::AllocatePackBuf and FluidSystem::UnpackParticles.

- Fixes a printf format warning

- Fixes several warnings related to double-precision literals being converted to floats

- Fixes a bug where enabling/re-enabling the color channel would cause a vector
subscript out of bounds error. However, it looks like it's still possible to get
the sample to crash by mashing keys (I think this may be GPU synchronization related).

- Fixes a warning from valgrind about a conditional on an uninitialized part of
memory by initializing mbDebug to false in the allocator constructor. (GVDB sets
mbDebug immediately after initializing the allocator.)

- Fixes a few bugs in GVDB where memory was being allocated using malloc, but
then freed using delete. In most cases, using new[] and delete[] is shorter.

- Fixes cast warnings and replaces malloc with new[] in main_win.cpp and nv_gui.cpp.