Adds gNanoVDB, a new sample showing how to create a NanoVDB volume from a GVDB volume in a high-performance way almost entirely on the GPU. In the future, this functionality will be added to the core library. (Note that although converting the Stanford Dragon (imported via LoadVDB) to a ~200 MB NanoVDB data structure takes 80 milliseconds on a GTX 1060, I think it's possible to optimize this further.)
In order to build this sample, you'll need to enable GVDB_BUILD_OPENVDB, and set
GVDB_NANOVDB_INCLUDE_DIR to the folder in OpenVDB containing a folder named
`nanovdb` with several header files and a `util` directory inside. To get NanoVDB,
you'll need to checkout the `feature/nanovdb` branch of OpenVDB, at
https://github.com/AcademySoftwareFoundation/openvdb/tree/feature/nanovdb
This might seem complicated - however, since NanoVDB is a header-only library,
you don't have to build any dependencies to run this sample!

Also, adds several functions to GVDB to facilitate retrieving data from external
applications more easily from the CPU (gNanoVDB doesn't use most of these, but
the CPU reference implementation does)

- AtlasRetrieveBrickXYZ, an API-facing form of mPool->AtlasRetrieveTexXYZ.
(Note that calling this function in a loop will add lots of synchronization
overhead; a more efficient data-driven-style version may be added soon).

- GetChannelType returns the type of the given channel.

- getChildRefAtBit takes a node, and returns the bth child reference. (It's
essentially two-thirds of getChildAtBit).

- FreeData allows you to free data that you've allocated using AllocData.

Finally, removes a warning during CMake configuration when building with support
for OpenVDB. This was due to how for compatibility with past renamings of the
OpenVDB debug DLL, we look for both openvdb.dll and openvdb_d.dll - but usually,
one of these does not exist, and so generates a warning when we look for it. We
now produce a warning only if openvdb.dll couldn't be found.